refactor(cntt-import): add manifest-driven batch import and cleanup
Implement manifest-driven batch import in the async import pipeline,
moving manifest schema, validation, and job expansion into the core
library. Replace tool-specific manifest/model code with unified
ImportManifest APIs. Update all CLI commands and tests to use new
request builders and manifest loading. Add LooseCookedIndexRegistry for
safe concurrent index updates. Improve job submission, shutdown, and
resource management. Extend texture import with multi-source and new
tuning options. Update manifests, schema, and documentation for new
field names and structure. Clean up code, improve logging, and update
build system for new dependencies.

BREAKING CHANGE: Manifest schema and CLI options have changed. Old
manifests and tool-specific manifest/model code are no longer supported.

+++ projects/Oxygen.Engine/src/Oxygen/Content/Import/Internal/Utils/AssetKeyUtils.h
@@ -19,8 +19,8 @@ namespace oxygen::content::import::util {
 [[nodiscard]] inline auto MakeDeterministicAssetKey(
   std::string_view virtual_path) -> data::AssetKey
 {
-  const auto bytes = std::as_bytes(
-    std::span(virtual_path.data(), static_cast<size_t>(virtual_path.size())));
+  const auto bytes
+    = std::as_bytes(std::span(virtual_path.data(), virtual_path.size()));
   const auto digest = base::ComputeSha256(bytes);
 
   data::AssetKey key {};
